+++ Package.swift
@@ -22,7 +22,6 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/StanfordSpezi/Spezi", .upToNextMinor(from: "0.7.0")),
-        // .package(path: "../SpeziViews")
         .package(url: "https://github.com/StanfordSpezi/SpeziViews", branch: "feat/view-state-enhancements")
     ],
     targets: [
