+++ Sources/SpeziOnboarding/Resources/en.lproj/Localizable.strings
@@ -10,6 +10,16 @@
 "CONSENT_ACTION" = "I Consent";
 
 
+// MARK: Consent Document
+"MARKDOWN_LOADING_ERROR" = "Could not load and parse the document.";
+"EXPORTED_TAG" = "Exported";
+"CONSENT_TITLE" = "Spezi Consent";
+"FILE_NAME_EXPORTED_CONSENT_FORM" = "Signed Consent Form";
+"CONSENT_EXPORT_ERROR_DESCRIPTION" = "Failed to export the signed consent form.";
+"CONSENT_EXPORT_ERROR_RECOVERY_SUGGESTION" = "Please try again or restart the application.";
+"CONSENT_EXPORT_ERROR_FAILURE_REASON" = "The system wasn't able to reserve the necessary memory for rendering the consent form.";
+"CONSENT_VIEW_TITLE" = "Consent";
+
 // MARK: Sequential Onboarding
 "SEQUENTIAL_ONBOARDING_NEXT" = "Next";
 
@@ -29,9 +39,3 @@
 
 // MARK: Illegal Onboarding Step
 "ILLEGAL_ONBOARDING_STEP" = "SpeziOnboarding: Illegal Onboarding Step";
-
-// MARK: Consent Form Export
-"MARKDOWN_LOADING_ERROR" = "Could not load and parse the document.";
-"EXPORTED_TAG" = "Exported";
-"CONSENT_TITLE" = "Spezi Consent";
-"FILE_NAME_EXPORTED_CONSENT_FORM" = "Signed Consent Form";
